+++ sdk/keyvault/azure-keyvault-certificates/tests/test_merge_certificate_async.py
@@ -3,13 +3,13 @@
 # Licensed under the MIT License.
 # ------------------------------------
 import base64
+import platform
 import os
 
 import pytest
 from azure.keyvault.certificates import CertificatePolicy, WellKnownIssuerNames
 from devtools_testutils import set_bodiless_matcher
 from devtools_testutils.aio import recorded_by_proxy_async
-from OpenSSL import crypto
 
 from _async_test_case import AsyncCertificatesClientPreparer
 from _test_case import get_decorator
@@ -19,11 +19,13 @@ all_api_versions = get_decorator(is_async=True)
 
 
 class TestMergeCertificate(KeyVaultTestCase):
+    @pytest.mark.skipif(platform.python_implementation() == "PyPy", reason="PyPy has issues with OpenSSL")
     @pytest.mark.asyncio
     @pytest.mark.parametrize("api_version", all_api_versions)
     @AsyncCertificatesClientPreparer(logging_enable=True)
     @recorded_by_proxy_async
     async def test_merge_certificate(self, client, **kwargs):
+        from OpenSSL import crypto
         set_bodiless_matcher()
         cert_name = self.get_resource_name("mergeCertificate")
         cert_policy = CertificatePolicy(